requirements.txt
uvicorn
fastapi
pydantic
python-multipart
environs
passlib[bcrypt]
python-jose[cryptography]
databases[postgresql]
sqlalchemy
psycopg2-binary